// sdk/src/runtime/generated/runtime/v1/external_agent.ts
export interface ExternalAgentActionScope {
    /**
     * @generated from protobuf field: string action_id = 1
     */
    actionId: string;
    /**
     * @generated from protobuf field: repeated string ops = 2
     */
    ops: string[];
}
/**
 * @generated from protobuf message nimi.runtime.v1.ExternalAgentGatewayStatusRequest
 */
export interface ExternalAgentGatewayStatusRequest {
}
/**
 * @generated from protobuf message nimi.runtime.v1.ExternalAgentGatewayStatusResponse
 */
export interface ExternalAgentGatewayStatusResponse {
    /**
     * @generated from protobuf field: bool enabled = 1
     */
    enabled: boolean;
    /**
     * @generated from protobuf field: string bind_address = 2
     */
    bindAddress: string;
    /**
     * @generated from protobuf field: string issuer = 3
     */
    issuer: string;
    /**
     * @generated from protobuf field: int32 action_count = 4
     */
    actionCount: number;
    /**
     * @generated from protobuf field: string status = 5
     */
    status: string;
    /**
     * @generated from protobuf field: string reason_code = 6
     */
    reasonCode: string;
}
// ...
export interface ExternalAgentIssueTokenResponse {
    /**
     * @generated from protobuf field: string token = 1
     */
    token: string;
    /**
     * @generated from protobuf field: string token_id = 2
     */
    tokenId: string;
    /**
     * @generated from protobuf field: string principal_id = 3
     */
    principalId: string;
    /**
     * @generated from protobuf field: string mode = 4
     */
    mode: string;
    /**
     * @generated from protobuf field: string subject_account_id = 5
     */
    subjectAccountId: string;
    /**
     * @generated from protobuf field: repeated string actions = 6
     */
    actions: string[];
    /**
     * @generated from protobuf field: repeated nimi.runtime.v1.ExternalAgentActionScope scopes = 7
     */
    scopes: ExternalAgentActionScope[];
    /**
     * @generated from protobuf field: google.protobuf.Timestamp issued_at = 8
     */
    issuedAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp expires_at = 9
     */
    expiresAt?: Timestamp;
    /**
     * @generated from protobuf field: string revoked_at = 10
     */
    revokedAt: string;
    /**
     * @generated from protobuf field: string issuer = 11
     */
    issuer: string;
}
/**
 * @generated from protobuf message nimi.runtime.v1.ExternalAgentRevokeTokenRequest
 */
export interface ExternalAgentRevokeTokenRequest {
    /**
     * @generated from protobuf field: string token_id = 1
     */
    tokenId: string;
}
/**
 * @generated from protobuf message nimi.runtime.v1.ExternalAgentTokenRecord
 */
export interface ExternalAgentTokenRecord {
    /**
     * @generated from protobuf field: string token_id = 1
     */
    tokenId: string;
    /**
     * @generated from protobuf field: string principal_id = 2
     */
    principalId: string;
    /**
     * @generated from protobuf field: string mode = 3
     */
    mode: string;
    /**
     * @generated from protobuf field: string subject_account_id = 4
     */
    subjectAccountId: string;
    /**
     * @generated from protobuf field: repeated string actions = 5
     */
    actions: string[];
    /**
     * @generated from protobuf field: repeated nimi.runtime.v1.ExternalAgentActionScope scopes = 6
     */
    scopes: ExternalAgentActionScope[];
    /**
     * @generated from protobuf field: google.protobuf.Timestamp issued_at = 7
     */
    issuedAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp expires_at = 8
     */
    expiresAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp revoked_at = 9
     */
    revokedAt?: Timestamp;
    /**
     * @generated from protobuf field: string issuer = 10
     */
    issuer: string;
}
// ...
class ExternalAgentActionScope$Type extends MessageType<ExternalAgentActionScope> {
    constructor() {
        super("nimi.runtime.v1.ExternalAgentActionScope", [
            { no: 1, name: "action_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "ops", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ExternalAgentActionScope>): ExternalAgentActionScope {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.actionId = "";
        message.ops = [];
        if (value !== undefined)
            reflectionMergePartial<ExternalAgentActionScope>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ExternalAgentActionScope): ExternalAgentActionScope {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string action_id */ 1:
                    message.actionId = reader.string();
                    break;
                case /* repeated string ops */ 2:
                    message.ops.push(reader.string());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ExternalAgentActionScope, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string action_id = 1; */
        if (message.actionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.actionId);
        /* repeated string ops = 2; */
        for (let i = 0; i < message.ops.length; i++)
            writer.tag(2, WireType.LengthDelimited).string(message.ops[i]);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.ExternalAgentActionScope
 */
export const ExternalAgentActionScope = new ExternalAgentActionScope$Type();
// ...
class ExternalAgentGatewayStatusResponse$Type extends MessageType<ExternalAgentGatewayStatusResponse> {
    constructor() {
        super("nimi.runtime.v1.ExternalAgentGatewayStatusResponse", [
            { no: 1, name: "enabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 2, name: "bind_address", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "issuer", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "action_count", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 5, name: "status", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "reason_code", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ExternalAgentGatewayStatusResponse>): ExternalAgentGatewayStatusResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.enabled = false;
        message.bindAddress = "";
        message.issuer = "";
        message.actionCount = 0;
        message.status = "";
        message.reasonCode = "";
        if (value !== undefined)
            reflectionMergePartial<ExternalAgentGatewayStatusResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ExternalAgentGatewayStatusResponse): ExternalAgentGatewayStatusResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* bool enabled */ 1:
                    message.enabled = reader.bool();
                    break;
                case /* string bind_address */ 2:
                    message.bindAddress = reader.string();
                    break;
                case /* string issuer */ 3:
                    message.issuer = reader.string();
                    break;
                case /* int32 action_count */ 4:
                    message.actionCount = reader.int32();
                    break;
                case /* string status */ 5:
                    message.status = reader.string();
                    break;
                case /* string reason_code */ 6:
                    message.reasonCode = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ExternalAgentGatewayStatusResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* bool enabled = 1; */
        if (message.enabled !== false)
            writer.tag(1, WireType.Varint).bool(message.enabled);
        /* string bind_address = 2; */
        if (message.bindAddress !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.bindAddress);
        /* string issuer = 3; */
        if (message.issuer !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.issuer);
        /* int32 action_count = 4; */
        if (message.actionCount !== 0)
            writer.tag(4, WireType.Varint).int32(message.actionCount);
        /* string status = 5; */
        if (message.status !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.status);
        /* string reason_code = 6; */
        if (message.reasonCode !== "")
            writer.tag(6, WireType.LengthDelimited).string(message.reasonCode);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.ExternalAgentGatewayStatusResponse
 */
export const ExternalAgentGatewayStatusResponse = new ExternalAgentGatewayStatusResponse$Type();
// ...
export const ExternalAgentIssueTokenResponse = new ExternalAgentIssueTokenResponse$Type();
// ...
export const ExternalAgentTokenRecord = new ExternalAgentTokenRecord$Type();